move fix function to its own subcommand
